+++ Assembly-CSharp/Global/PLAYER_BASE.cs
@@ -2,9 +2,9 @@
 
 public class PLAYER_BASE
 {
-	public Int16 max_hp;
+	public UInt32 max_hp;
 
-	public Int16 max_mp;
+	public UInt32 max_mp;
 
 	public Byte dex;
