+++ Assembly-CSharp/Global/ff9/FF9PLAY_SKILL.cs
@@ -8,13 +8,13 @@ public class FF9PLAY_SKILL
 		this.weapon = new UInt16[5];
 	}
 
-	public UInt16 cur_hp;
+	public UInt32 cur_hp;
 
-	public UInt16 cur_mp;
+	public UInt32 cur_mp;
 
-	public UInt16 max_hp;
+	public UInt32 max_hp;
 
-	public UInt16 max_mp;
+	public UInt32 max_mp;
 
 	public Byte[] Base;
