+++ Memoria.Scripts/Sources/Battle/0029_DifferentCasterHpAttackScript.cs
@@ -20,7 +20,7 @@ namespace Memoria.Scripts.Battle
         public void Perform()
         {
             _v.Target.Flags |= CalcFlag.HpAlteration;
-            _v.Target.HpDamage = (Int16)(_v.Caster.MaximumHp - _v.Caster.CurrentHp);
+            _v.Target.HpDamage = (Int32)(_v.Caster.MaximumHp - _v.Caster.CurrentHp);
         }
     }
 }
