+++ Memoria.Scripts/Sources/Battle/0042_ThrowScript.cs
@@ -25,16 +25,16 @@ namespace Memoria.Scripts.Battle
 
             _v.Caster.SetLowPhisicalAttack();
             _v.Target.SetPhisicalDefense();
-            _v.Context.AttackPower = (Int16)(_v.Command.Weapon.Power << 1);
+            _v.Context.AttackPower = _v.Command.Weapon.Power << 1;
 
             if (_v.Caster.HasSupportAbility(SupportAbility1.PowerThrow))
-                _v.Context.Attack = (Int16)(_v.Context.Attack * 3 >> 1);
+                _v.Context.Attack = _v.Context.Attack * 3 >> 1;
 
             _v.Caster.PhysicalPenaltyAndBonusAttack();
             _v.Target.PhysicalPenaltyAndBonusAttack();
 
             if (_v.Target.HasCategory(EnemyCategory.Flight))
-                _v.Context.Attack = (Int16)(_v.Context.Attack * 3 >> 1);
+                _v.Context.Attack = _v.Context.Attack * 3 >> 1;
 
             _v.CalcPhysicalHpDamage();
         }
