+++ Memoria.Scripts/Sources/Battle/0087_OdinScript.cs
@@ -31,8 +31,8 @@ namespace Memoria.Scripts.Battle
             if ((_v.Context.Flags & (BattleCalcFlags.Guard | BattleCalcFlags.Miss)) != 0 && _v.Caster.HasSupportAbility(SupportAbility2.OdinSword))
             {
                 _v.Context.Flags &= (BattleCalcFlags)65470;
-                _v.Context.Attack = (Int16)(_v.Caster.Magic + GameRandom.Next16() % (1 + (_v.Caster.Level + _v.Caster.Magic >> 3)));
-                _v.Context.AttackPower = (Int16)(_v.Command.Power + 100 - ff9item.FF9Item_GetCount((Int16)GemItem.Ore));
+                _v.Context.Attack = _v.Caster.Magic + GameRandom.Next16() % (1 + (_v.Caster.Level + _v.Caster.Magic >> 3));
+                _v.Context.AttackPower = _v.Command.Power + 100 - ff9item.FF9Item_GetCount((Int16)GemItem.Ore);
                 _v.Context.DefensePower = _v.Target.MagicDefence;
                 _v.TargetCommand.CalcHpDamage();
             }
